Fix the logging api, bump to next major revision due to breaking changes
Shouldn't log every time an error is created. Only log the error at the
top level. Changes made reflect this. The old api still exists but the
log.error call should now only be made at the top level (so usually in
the framework not by apps - breaking).
Add an explicit init for the logging to initialise logentries. Breaking.
Add content to the readme to reflect the changes.